refactor(agent): the iteration-limit summary lives in its own module
`agent/chat_completion_helpers.py` is 3,569 lines, well past the ~2,000-line
split signal in AGENTS.md, and the next change to this area adds a fourth
provider to it. Split first, in its own commit, as that rule asks.

The unit is self-contained: eleven symbols, contiguous, and every one of the
ten private ones appears exactly twice in the owner — its definition and its
single use inside this unit. So this is a move, not an untangling.

Two helpers stay behind because the main loop uses them too
(`_provider_preferences_for_agent` at :1477, `_merge_nous_portal_messages_extra_body`
at :1403). They are imported inside the two functions that need them, matching
this unit's existing idiom for `relay_llm`, `MAX_ITERATIONS_SUMMARY_REQUEST` and
`get_provider_profile`, which keeps the dependency one-way and the new module
free of any module-level import of its former owner.

`handle_max_iterations` is re-exported from `chat_completion_helpers` and has to
stay reachable there: `run_agent.py:1329` binds it through
`agent.lazy_forward.forward("agent.chat_completion_helpers", "handle_max_iterations")`,
which resolves the attribute on that module at every call so tests can patch it
("Targets are resolved on every call" — lazy_forward's own docstring), and six
test modules patch or import it by that path.

No behaviour change. 257 tests covering the seam and every module that touches
the moved symbols pass unmodified, which is the proof the move is a move:
tests/agent/{test_api_content_sidecar,test_turn_finalizer_cleanup_guard,
test_turn_finalizer_final_response_persistence,test_turn_finalizer_iteration_limit_exit,
test_aux_progress_streaming,test_ghost_skill_pruning,test_micro_compaction}.py,
tests/run_agent/{test_verification_continuation_budget,
test_native_compaction_summary_retention,test_run_agent_codex_responses}.py and
test_run_agent.py -k "max_iteration or summary or iteration".
